Move tests from o.p.jaxrs to o.p.jaxrs.tests
... and rename two base classes to indicate those use the
DatabaseAdapter.
