+++ servers/jax-rs-tests/src/test/java/org/projectnessie/jaxrs/tests/AbstractTestDatabaseAdapterRest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 Dremio
+ * Copyright (C) 2022 Dremio
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.projectnessie.jaxrs;
+package org.projectnessie.jaxrs.tests;
 
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.extension.RegisterExtension;
@@ -24,7 +24,7 @@ import org.projectnessie.versioned.persist.tests.extension.NessieDbAdapter;
 
 /** See {@link AbstractTestRest} for details about and reason for the inheritance model. */
 @ExtendWith(DatabaseAdapterExtension.class)
-abstract class AbstractTestJerseyRest extends AbstractRestSecurityContext {
+abstract class AbstractTestDatabaseAdapterRest extends AbstractRestSecurityContext {
 
   @NessieDbAdapter static DatabaseAdapter databaseAdapter;
