+++ servers/jax-rs-tests/src/test/java/org/projectnessie/jaxrs/tests/AbstractTestDatabaseAdapterResteasy.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 Dremio
+ * Copyright (C) 2022 Dremio
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.projectnessie.jaxrs;
+package org.projectnessie.jaxrs.tests;
 
 import io.restassured.RestAssured;
 import io.restassured.builder.RequestSpecBuilder;
@@ -29,7 +29,7 @@ import org.projectnessie.versioned.persist.tests.extension.DatabaseAdapterExtens
 import org.projectnessie.versioned.persist.tests.extension.NessieDbAdapter;
 
 @ExtendWith(DatabaseAdapterExtension.class)
-abstract class AbstractTestJerseyResteasy extends AbstractResteasyTest {
+abstract class AbstractTestDatabaseAdapterResteasy extends AbstractResteasyTest {
 
   @NessieDbAdapter static DatabaseAdapter databaseAdapter;
