+++ servers/jax-rs-tests/src/test/java/org/projectnessie/jaxrs/tests/ITJerseyRestRocks.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 Dremio
+ * Copyright (C) 2022 Dremio
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.projectnessie.jaxrs;
+package org.projectnessie.jaxrs.tests;
 
 import org.projectnessie.versioned.persist.rocks.RocksDatabaseAdapterFactory;
 import org.projectnessie.versioned.persist.rocks.RocksTestConnectionProviderSource;
@@ -22,4 +22,4 @@ import org.projectnessie.versioned.persist.tests.extension.NessieExternalDatabas
 
 @NessieDbAdapterName(RocksDatabaseAdapterFactory.NAME)
 @NessieExternalDatabase(RocksTestConnectionProviderSource.class)
-class ITJerseyRestRocks extends AbstractTestJerseyRest {}
+class ITJerseyRestRocks extends AbstractTestDatabaseAdapterRest {}
