+++ servers/jax-rs-tests/src/test/java/org/projectnessie/jaxrs/tests/TestJerseyRestNaiveClientInMemory.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 Dremio
+ * Copyright (C) 2022 Dremio
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.projectnessie.jaxrs;
+package org.projectnessie.jaxrs.tests;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.projectnessie.client.http.impl.HttpUtils.HEADER_ACCEPT;
@@ -44,7 +44,7 @@ import org.projectnessie.versioned.persist.tests.extension.NessieExternalDatabas
  */
 @NessieDbAdapterName(InmemoryDatabaseAdapterFactory.NAME)
 @NessieExternalDatabase(InmemoryTestConnectionProviderSource.class)
-class TestJerseyRestNaiveClientInMemory extends AbstractTestJerseyRest
+class TestJerseyRestNaiveClientInMemory extends AbstractTestDatabaseAdapterRest
     implements NessieClientCustomizer {
 
   private boolean headersProcessed;
